Translations loaded too early error
Notice: Function _load_textdomain_just_in_time was called incorrectly.
Translation loading for the urvanov-syntax-highlighter domain was
triggered too early. This is usually an indicator for some code in the
plugin or theme running too early. Translations should be loaded at the
init action or later. Please see Debugging in WordPress for more
information. (This message was added in version 6.7.0.) in
/members/cimabue/blog/wp-includes/functions.php on line 6121

https://wordpress.org/support/topic/translations-loaded-too-early-error/